create useVisibility hook
